fix: satisfy gui lint after merge
